Improve annotation commontator::comment association

+++ app/controllers/annotations_controller.rb
@@ -1,55 +1,53 @@
 class AnnotationsController < ApplicationController
 
-  before_action :update_comments, only: [:edit, :update_annotations]
-
   authorize_resource
 
   def new
     @annotation = Annotation.new(category: :note, color: Annotation.colors[1])
+
     @total_seconds = params[:total_seconds]
-    @medium_id = params[:mediumId]
-    @posted = (@annotation.public_comment_id != nil)
+    @medium_id = params[:medium_id]
+    @posted = false
 
     render :edit
   end
 
   def edit
-    @annotation = Annotation.find(params[:annotationId])
+    @annotation = Annotation.find(params[:annotation_id])
+
     # only allow editing, if the current user created the annotation
     if @annotation.user_id != current_user.id
       render json: false
       return
     end
+
     @total_seconds = @annotation.timestamp.total_seconds
     @medium_id = @annotation.medium_id
     @posted = (@annotation.public_comment_id != nil)
-  end
 
-  def update
-    @annotation = Annotation.find(params[:id])
-    comment_id = @annotation.public_comment_id
-    if (comment_id == nil)
-      @annotation.public_comment_id = post_comment
-    else
-      medium = Medium.find_by_id(params[:annotation][:medium_id])
-      total_seconds = params[:annotation][:total_seconds]
-      comment = params[:annotation][:comment]
-      commontator_comment = Commontator::Comment.find_by(id: comment_id)
-      commontator_comment.update(editor: current_user,
-                                 body: comment_html(medium, total_seconds, comment))
-    end
-    @annotation.update(annotation_params)
+    # if this annotation has an associated commontator comment,
+    # we have to call the "get_comment" method in order to get
+    # the text.
+    @annotation.comment = @annotation.get_comment
   end
 
   def create
     @annotation = Annotation.new(annotation_params)
     @annotation.user_id = current_user.id
-    @total_seconds = params[:annotation][:total_seconds]
+    @total_seconds = annotation_auxiliary_params[:total_seconds]
     @annotation.timestamp = TimeStamp.new(total_seconds: @total_seconds)
-    @annotation.public_comment_id = post_comment
+    @annotation.public_comment_id = post_comment(@annotation)
 
     @annotation.save
     render :update
+  end
+
+  def update
+    @annotation = Annotation.find(params[:id])
+    @annotation.update(annotation_params)
+
+    @annotation.public_comment_id = post_comment(@annotation)
+    @annotation.save
   end
 
   def show
@@ -57,10 +55,19 @@ class AnnotationsController < ApplicationController
   end
 
   def destroy
-    annotation = Annotation.find(params[:annotationId])
-    comment = Commontator::Comment.find_by(id: annotation.public_comment_id)
-    comment.update(deleted_at: DateTime.now)
+    annotation = Annotation.find(params[:annotation_id])
+
+    # only the owner of the annotation is allowed to delete it
+    return unless annotation.user == current_user
+
+    # delete associated commontator comment
+    unless annotation.public_comment_id.nil?
+      commontator_comment = Commontator::Comment.find_by(id: annotation.public_comment_id)
+      commontator_comment.update(deleted_at: DateTime.now)
+    end
+
     annotation.destroy
+
     render json: []
   end
 
@@ -78,14 +85,21 @@ class AnnotationsController < ApplicationController
                                      user: current_user)
     end
 
+    # If annotation is associated to a comment,
+    # the field "comment" is empty -> get it from the commontator comment
+    annotations.each do |a|
+      a.comment = a.get_comment
+    end
+
     # Convert to JSON (for easier hash operations)
     annotations = annotations.as_json
 
     # Filter attributes and add boolean "belongs_to_current_user".
     annotations.each do |a|
+      public_comment_id = a['public_comment_id']
       a['belongs_to_current_user'] = (current_user.id == a['user_id'])
-      a.slice!('category', 'color', 'comment', 'id',
-               'belongs_to_current_user', 'timestamp', 'subcategory')
+      a.slice!('belongs_to_current_user', 'category', 'color', 'comment',
+               'id', 'subcategory', 'timestamp')
     end
 
     render json: annotations
@@ -117,44 +131,43 @@ class AnnotationsController < ApplicationController
 
     def annotation_params
       params.require(:annotation).permit(
-        :category, :color, :comment, :medium_id, :subcategory, :visible_for_teacher)
+        :category, :color, :medium_id, :subcategory, :visible_for_teacher
+      )
     end
 
-    def post_comment
-      if params[:annotation][:post_as_comment] == "1"
-        medium = Medium.find_by_id(params[:annotation][:medium_id])
-        total_seconds = params[:annotation][:total_seconds]
-        commontator_thread = medium.commontator_thread
-        commontator_comment = Commontator::Comment.new(
-          thread: commontator_thread,
+    def annotation_auxiliary_params
+      params.require(:annotation).permit(
+        :total_seconds, :post_as_comment, :comment
+      )
+    end
+
+    def post_comment(annotation)
+      public_comment_id = annotation.public_comment_id
+      aux_params = annotation_auxiliary_params
+
+      # return if checkbox "post_as_comment" is not checked and if there is no comment to update
+      return if aux_params[:post_as_comment] != "1" && public_comment_id.nil?
+
+      comment = aux_params[:comment]
+
+      if public_comment_id.nil? # comment doesn't exist yet -> create one
+        medium = annotation.medium
+        commontator_comment = Commontator::Comment.create(
+          thread: medium.commontator_thread,
           creator: current_user,
-          body: comment_html = comment_html(medium,
-                                            total_seconds,
-                                            annotation_params[:comment])
+          body: comment
         )
-        commontator_comment.save
-        return commontator_comment.id
+        commontator_comment.annotation = annotation
+      else # comment already exists -> update it
+        commontator_comment = Commontator::Comment.find_by(id: public_comment_id)
+        commontator_comment.update(editor: current_user,
+                                     body: comment)
       end
-    end
 
-    def comment_html(medium, total_seconds, comment)
-      timestamp = TimeStamp.new(total_seconds: total_seconds).hms_colon_string
-      return comment + "\n" +
-             'Thymestamp: <a href="' + play_medium_path(medium) +
-             '?time=' + total_seconds + '">' + timestamp + '</a>'
-    end
+      # delete comment as it is already saved in the commontator comment model
+      annotation.comment = nil
 
-    def update_comments
-      Annotation.where(medium: params[:mediumId]).each do |annotation|
-        # find comment
-        unless annotation.public_comment.nil?
-          comment = annotation.public_comment.body
-
-          # remove "Thymestamp: H:MM:SS" at the end of the string
-          index = comment.rindex("\nThymestamp")
-          annotation.update(comment: comment[0 .. index - 1]) unless index == nil
-        end
-      end
+      return commontator_comment.id
     end
 
 end
